+++ src/database/migrations/1.0.0/1699131121605-DHCP_Service.ts
@@ -1,0 +1,11 @@
+import {MigrationInterface, QueryRunner} from "typeorm";
+
+export class DHCPService1699131121605 implements MigrationInterface {
+
+    public async up(queryRunner: QueryRunner): Promise<void> {
+    }
+
+    public async down(queryRunner: QueryRunner): Promise<void> {
+    }
+
+}