+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -
 ### Added
+- Database migration for the DHCP service management.
 - In the firewalls/clusters tree new sort of nodes under a firewall or cluster node. 
 - Migrations are added to create the System nodes and their children: DHCP, Keepalived, and HAProxy.